web: Move the actual creation of panic error/actions into panic.tsx
